Move all selectors to ./selectors Move all actions to ./actions
